requirements.txt
adjustText==1.2.0
affine==2.4.0
aiobotocore==2.11.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioitertools==0.11.0
aiosignal==1.3.1
alabaster==1.0.0
anndata==0.10.8
annotated-types==0.7.0
app-model==0.2.8
appdirs==1.4.4
array_api_compat==1.8
asciitree==0.3.3
astropy==6.1.3
astropy-iers-data==0.2024.9.16.0.32.21
asttokens==2.4.1
attrs==24.2.0
babel==2.16.0
-e git+ssh://git@github.com/ckmah/bento-tools.git@ced6bf777738cb681b24a296de3e7e0a6c6ec88d#egg=bento_tools
botocore==1.34.19
build==1.2.1
cachey==0.2.1
cellpose==3.0.11
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
colorcet==3.1.0
comm==0.2.2
ConfigArgParse==1.7
connection_pool==0.0.3
contourpy==1.2.1
cycler==0.12.1
dask==2024.9.0
dask-expr==1.1.14
dask-image==2023.8.1
datashader==0.16.3
datrie==0.8.2
debugpy==1.8.5
decorator==5.1.1
decoupler==1.8.0
distributed==2024.9.0
docstring_parser==0.16
docutils==0.21.2
dpath==2.2.0
emoji==2.13.0
executing==2.0.1
fasteners==0.19
fastjsonschema==2.20.0
fastremap==1.15.0
fcsparser==0.2.8
filelock==3.15.4
flake8==7.1.1
flexcache==0.3
flexparser==0.3.1
fonttools==4.53.1
freetype-py==2.4.0
frozenlist==1.4.1
fsspec==2023.6.0
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
h5py==3.11.0
HeapDict==1.0.1
hsluv==5.0.4
humanfriendly==10.0
idna==3.8
igraph==0.11.6
imagecodecs==2024.6.1
imageio==2.35.1
imagesize==1.4.1
importlib_metadata==8.4.0
in-n-out==0.2.1
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
kneed==0.8.5
lamin_utils==0.13.4
lazy_loader==0.4
legacy-api-wrap==1.4
leidenalg==0.10.2
llvmlite==0.43.0
locket==1.0.0
loguru==0.7.2
magicgui==0.9.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
matplotlib-scalebar==0.8.1
mccabe==0.7.0
mdurl==0.1.2
MiniSom==2.3.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multipledispatch==1.0.0
multiscale_spatial_image==1.0.1
napari==0.4.19.post1
napari-console==0.0.9
napari-matplotlib==2.0.3
napari-plugin-engine==0.2.0
napari-spatialdata==0.5.2.post1
napari-svg==0.2.0
natsort==8.4.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
npe2==0.7.7
numba==0.60.0
numcodecs==0.13.0
numpy==1.26.4
numpydoc==1.8.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
ome-zarr==0.9.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
param==2.1.1
parso==0.8.4
partd==1.4.2
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
PIMS==0.7
Pint==0.24.3
plac==1.4.3
platformdirs==4.2.2
pooch==1.8.2
prompt_toolkit==3.0.47
psutil==6.0.0
psygnal==0.11.1
ptyprocess==0.7.0
PuLP==2.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycodestyle==2.12.1
pyconify==0.1.6
pyct==0.5.0
pydantic==2.8.2
pydantic-compat==0.1.2
pydantic_core==2.20.1
pyerfa==2.0.1.4
pyflakes==3.2.0
pygeos==0.14
Pygments==2.18.0
pynndescent==0.5.13
pyogrio==0.9.0
PyOpenGL==3.1.7
pyparsing==3.1.4
pyproj==3.6.1
pyproject_hooks==1.1.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
pyqtgraph==0.13.7
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.5.2
QtPy==2.4.1
rasterio==1.3.10
readfcs==1.1.8
referencing==0.35.1
requests==2.32.3
reretry==0.11.8
rich==13.8.0
rioxarray==0.17.0
roifile==2024.5.24
rpds-py==0.20.0
Rtree==1.3.0
s3fs==0.6.0
scanpy==1.10.2
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.10.1
seaborn==0.13.2
session_info==1.0.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
slicerator==1.1.0
smart-open==7.0.4
smmap==5.0.1
snakemake==7.32.4
snowballstemmer==2.2.0
snuggs==1.4.7
sopa==1.1.4
sortedcontainers==2.4.0
sparse==0.15.4
spatial_image==1.1.0
spatialdata==0.2.2
-e git+ssh://git@github.com/ckmah/spatialdata-io.git@15f8f68772d3ee394e86e7a7e9655233c666bd63#egg=spatialdata_io
spatialdata-plot==0.2.4
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
statsmodels==0.14.2
stdlib-list==0.10.0
stopit==1.1.2
superqt==0.6.7
sympy==1.13.2
tabulate==0.9.0
tblib==3.0.0
tensorly==0.8.1
texttable==1.7.0
threadpoolctl==3.5.0
throttler==1.2.2
tifffile==2024.8.24
tinycss2==1.3.0
tomli_w==1.0.0
toolz==0.12.1
toposort==1.10
torch==2.4.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
triton==3.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
umap-learn==0.5.6
UpSetPlot==0.9.0
urllib3==1.26.19
vispy==0.14.3
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.13
wrapt==1.16.0
xarray==2024.7.0
xarray-dataclasses==1.8.0
xarray-datatree==0.0.14
xarray-schema==0.0.3
xarray-spatial==0.4.0
xgboost==2.1.1
yarl==1.9.4
yte==1.5.4
zarr==2.18.2
zict==3.0.0
zipp==3.20.1
